[MISC] Move alignment result range into core.
