+++ include/seqan3/alignment/pairwise/align_pairwise.hpp
@@ -20,10 +20,10 @@
 #include <meta/meta.hpp>
 
 #include <seqan3/alignment/pairwise/alignment_configurator.hpp>
-#include <seqan3/alignment/pairwise/alignment_range.hpp>
 #include <seqan3/alignment/pairwise/alignment_result.hpp>
 #include <seqan3/alignment/pairwise/detail/concept.hpp>
 #include <seqan3/alignment/pairwise/detail/type_traits.hpp>
+#include <seqan3/core/algorithm/alignment_range.hpp>
 #include <seqan3/core/algorithm/detail/algorithm_executor_blocking.hpp>
 #include <seqan3/core/parallel/execution.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
