+++ include/seqan3/alignment/pairwise/all.hpp
@@ -16,7 +16,6 @@
 #include <seqan3/alignment/pairwise/align_result_selector.hpp>
 #include <seqan3/alignment/pairwise/alignment_algorithm.hpp>
 #include <seqan3/alignment/pairwise/alignment_configurator.hpp>
-#include <seqan3/alignment/pairwise/alignment_range.hpp>
 #include <seqan3/alignment/pairwise/alignment_result.hpp>
 #include <seqan3/alignment/pairwise/edit_distance_unbanded.hpp>
 #include <seqan3/alignment/pairwise/policy/all.hpp>
