+++ include/seqan3/core/algorithm/alignment_range.hpp
